+++ examples/kem.py
@@ -10,7 +10,7 @@ kems = oqs.get_enabled_kem_mechanisms()
 pprint(kems, compact=True)
 
 # Create client and server with sample KEM mechanisms
-kemalg = "Kyber512"
+kemalg = "ML-KEM-512"
 with oqs.KeyEncapsulation(kemalg) as client:
     with oqs.KeyEncapsulation(kemalg) as server:
         print("\nKey encapsulation details:")